LPS-54010 Convert ItemSelectorRendering into an interface to make API cleaner and easier to maintain.
